requirements.txt
APScheduler==3.6.3
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.9
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.1.5
Pillow==8.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
twilio==6.45.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.3.3
